requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
Werkzeug==3.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
